requirements.txt
jupyter
geopandas
netcdf4
pystac>=1.10
rasterio
shapely
xarray